Move Nitro attestation verification into type
